move functionality of DaceStrMaker to OirSDFGBuilder ctx
